+++ lib/impure/rdstdin.nim
@@ -13,15 +13,14 @@
 ## is used. This suffices because Windows' console already provides the
 ## wanted functionality.
 
-runnableExamples:
-  if false:
-    echo readLineFromStdin("Is Nim awesome? (Y/n): ")
-    var line: string
-    while true:
-      let ok = readLineFromStdin("How are you? ", line)
-      if not ok: break # ctrl-C or ctrl-D will cause a break
-      if line.len > 0: echo line
-    echo "exiting"
+runnableExamples("-r:off"):
+  echo readLineFromStdin("Is Nim awesome? (Y/n): ")
+  var line: string
+  while true:
+    let ok = readLineFromStdin("How are you? ", line)
+    if not ok: break # ctrl-C or ctrl-D will cause a break
+    if line.len > 0: echo line
+  echo "exiting"
 
 when defined(windows):
   proc readLineFromStdin*(prompt: string): string {.